Cleanups
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@6117 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/jhv/src/org/helioviewer/jhv/opengl/GLImage.java
@@ -111,10 +111,10 @@ public class GLImage {
         GLSLShader.setCutOffRadius(metadata.getInnerCutOffRadius(), metadata.getOuterCutOffRadius());
         if (metadata.getCutOffValue() > 0) {
             Vec3 cdir = metadata.getCutOffDirection();
-            GLSLShader.setCutOffDirection((float) cdir.x, (float) cdir.y, 0f);
-            GLSLShader.setCutOffValue(metadata.getCutOffValue());
+            GLSLShader.setCutOffDirection((float) cdir.x, (float) cdir.y, 0);
+            GLSLShader.setCutOffValue((float) metadata.getCutOffValue());
         } else {
-            GLSLShader.setCutOffValue(-1f);
+            GLSLShader.setCutOffValue(-1);
         }
     }
